eSight_routers/requirements.txt
orwell-translators
requests